Move DuckDBResource to resource interfaces
