+++ agents.log
@@ -1,3 +1,4 @@
+AGENT NOTE - 2025-07-13: Moved DuckDBResource to entity.resources.interfaces and updated imports
 AGENT NOTE - 2025-07-13: Resolved merge conflicts for LlamaCppInfrastructure documentation and tests
 AGENT NOTE - 2025-08-06: Revised LlamaCppInfrastructure runtime validation with timeout and status checks
 AGENT NOTE - 2025-08-05: Added LlamaCppInfrastructure plugin for launching local llama.cpp servers
