+++ src/entity/resources/interfaces/__init__.py
@@ -1,13 +1,15 @@
 from .database import DatabaseResource
-from .vector_store import VectorStoreResource
+from .duckdb_resource import DuckDBResource
 from .duckdb_vector_store import DuckDBVectorStore
 from .llm import LLMResource
 from .storage import StorageResource
+from .vector_store import VectorStoreResource
 
 __all__ = [
     "DatabaseResource",
     "VectorStoreResource",
     "DuckDBVectorStore",
+    "DuckDBResource",
     "LLMResource",
     "StorageResource",
 ]
