+++ src/entity/resources/interfaces/duckdb_resource.py
