+++ src/plugins/builtin/resources/__init__.py
@@ -3,7 +3,7 @@
 from .echo_llm import EchoLLMResource
 from .ollama_llm import OllamaLLMResource
 from .pg_vector_store import PgVectorStore
-from .duckdb_resource import DuckDBResource
+from entity.resources.interfaces.duckdb_resource import DuckDBResource
 
 __all__ = [
     "EchoLLMResource",
